move DockingInputsBase to WorkflowInputs
